verocta-ai-unified/backend/requirements.txt
# VEROCTA Financial Intelligence Platform - Core Dependencies
# Web framework
flask==3.1.1
werkzeug==3.1.3

# CORS and JWT authentication
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.1

# Data processing
pandas==2.3.1
numpy>=1.26.0

# AI integration
openai==1.98.0

# PDF generation
reportlab==4.4.3
matplotlib>=3.5.0

# Image processing
Pillow>=10.0.0

# Production server
gunicorn==23.0.0

# Date utilities
python-dateutil>=2.8.2
